fix(settlement): AI 리뷰 반영 - 수익 기록 트랜잭션 일관성 및 Long 타입 통일
- 🔴 ChapterPurchaseService: 수익 기록 try-catch 제거 → 실패 시 전체 롤백
- ⚠️ RevenueTransaction: creditAmount/platformFee/authorShare Integer → Long
- V8 마이그레이션: INTEGER → BIGINT, DTO/Service/테스트 동기화

+++ src/main/java/com/stolink/backend/domain/settlement/dto/RevenueTransactionResponse.java
@@ -12,10 +12,10 @@ public record RevenueTransactionResponse(
         UUID chapterId,
         UUID buyerUserId,
         RevenueTransactionType type,
-        Integer creditAmount,
+        Long creditAmount,
         Double platformFeeRate,
-        Integer platformFee,
-        Integer authorShare,
+        Long platformFee,
+        Long authorShare,
         UUID settlementId,
         LocalDateTime createdAt
 ) {
